Cleanup clippy warnings
