+++ tests/common/mod.rs
@@ -2,12 +2,10 @@ use std::path::{Path, PathBuf};
 use std::sync::OnceLock;
 use std::{collections::BTreeMap, fs};
 
-use libsla::{Address, AddressSpace, GhidraSleigh, OpCode, PcodeInstruction, Sleigh, VarnodeData};
-use pcode_ops::convert::PcodeValue;
+use libsla::{Address, GhidraSleigh, OpCode, PcodeInstruction, Sleigh, VarnodeData};
 use sym::{SymbolicBit, SymbolicBitVec, SymbolicByte};
 use symbolic_pcode::emulator::{self, ControlFlow, PcodeEmulator, StandardPcodeEmulator};
-use symbolic_pcode::mem::{GenericMemory, MemoryBranch, VarnodeDataStore};
-use symbolic_pcode::processor::{self, PcodeExecution, ProcessorResponseHandler};
+use symbolic_pcode::mem::{GenericMemory, VarnodeDataStore};
 
 static X86_64_SLA: OnceLock<PathBuf> = OnceLock::new();
 static AARCH_64_SLA: OnceLock<PathBuf> = OnceLock::new();
@@ -49,6 +47,7 @@ impl PcodeEmulator for TracingEmulator {
             | OpCode::CallIndirect
             | OpCode::Return => (),
             _ => {
+                /*
                 for instr_input in instruction.inputs.iter() {
                     let input_result = memory.read(instr_input);
                     let input_result = match input_result {
@@ -59,7 +58,6 @@ impl PcodeEmulator for TracingEmulator {
                         }
                     };
 
-                    /*
                     match u128::try_from(input_result) {
                         Ok(value) => {
                             println!(
@@ -74,8 +72,8 @@ impl PcodeEmulator for TracingEmulator {
                             println!("Input {instr_input} = Symbolic value @ {index}")
                         }
                     }
-                    */
                 }
+                    */
             }
         };
 
@@ -178,89 +176,6 @@ pub fn aarch64_sleigh() -> libsla::Result<GhidraSleigh> {
         .processor_spec(&processor_spec)?
         .build()?;
     Ok(sleigh)
-}
-
-pub fn initialize_libc_stack(memory: &mut Memory, sleigh: &impl Sleigh) {
-    // The stack for libc programs:
-    // * argc
-    // * argv - list must be terminated by NULL pointer
-    // * envp - list must be terminated by NULL pointer
-    // * auxv - list must be terminated by NULL pointer
-    let ram = sleigh
-        .address_space_by_name("ram")
-        .expect("failed to find ram");
-    let argc = VarnodeData {
-        address: Address {
-            offset: INITIAL_STACK,
-            address_space: ram.clone(),
-        },
-        size: 8,
-    };
-    memory
-        .write(&argc, SymbolicBitVec::constant(1, u64::BITS as usize))
-        .expect("failed to initialize argc on stack");
-
-    // The argv list must be terminated by null pointer. Setting program name to null AND
-    // terminating the list with NULL, whence 16 bytes
-    //
-    // musl has support for null program name:
-    // https://git.musl-libc.org/cgit/musl/tree/src/env/__libc_start_main.c
-    let argv = VarnodeData {
-        address: Address {
-            offset: argc.address.offset + argc.size as u64,
-            address_space: ram.clone(),
-        },
-        size: 16,
-    };
-    memory
-        .write(&argv, SymbolicBitVec::constant(0, (2 * u64::BITS) as usize))
-        .expect("failed to initialize argv");
-
-    let envp = VarnodeData {
-        address: Address {
-            offset: argv.address.offset + argv.size as u64,
-            address_space: ram.clone(),
-        },
-        size: 8,
-    };
-    memory
-        .write(&envp, SymbolicBitVec::constant(0, u64::BITS as usize))
-        .expect("failed to initialize envp");
-
-    // musl targets initialize the libc pagesize using aux[AT_PAGESZ]. For architectures without a
-    // hardcoded value the libc pagesize is used. For example x86-64 has a hardcoded value of 4096
-    // and so this is ignored. However for aarch64 there is no hardcoded value. This must be
-    // supplied otherwise a pagesize of 0 will be used in some cases.
-    let mut auxv = VarnodeData {
-        address: Address {
-            offset: envp.address.offset + envp.size as u64,
-            address_space: ram.clone(),
-        },
-        size: 8,
-    };
-
-    // The index for AT_PAGESZ
-    let at_pagesz = 6;
-    memory
-        .write(
-            &auxv,
-            SymbolicBitVec::constant(at_pagesz, u64::BITS as usize),
-        )
-        .expect("failed to write AT_PAGESZ into auxv");
-    auxv.address.offset += auxv.size as u64;
-
-    let page_size = 4096;
-    memory
-        .write(
-            &auxv,
-            SymbolicBitVec::constant(page_size, u64::BITS as usize),
-        )
-        .expect("failed to write page size into auxv");
-    auxv.address.offset += auxv.size as u64;
-
-    memory
-        .write(&auxv, SymbolicBitVec::constant(0, u64::BITS as usize))
-        .expect("failed to initialize auxv");
 }
 
 pub fn memory_with_image(
