+++ tests/main.rs
@@ -1,0 +1,4 @@
+mod common;
+mod hello_world;
+mod sleigh;
+mod x86_64_emulator;
